Add login to each Todo action to permit request

+++ Todo/tests/test_todo_endpoints.py
@@ -46,20 +46,21 @@ def auto_login_user(db, client):
     return make_auto_login
 
 
-def test_todo_list(db, client, create_todo):
+def test_todo_list(db, client, create_todo, auto_login_user):
     """
     Test that a list of todos are returned when GET request is made to '/api/todos'
     - @param client is a built-in fixture from pytest-django based off of django.test.client /
             which can be used for making requests to the app
     """
     
+    # login user
+    access_token, refresh_token = auto_login_user()
     # create a Todo object in db
     todo = create_todo(title="Learn how to use pytest", memo="Use the book 'Python Testing with Pytest'")
-    
     # get the url to for getting the todo list
     url = reverse('todo-list')
     # returns the response object from endpoint
-    response = client.get(url)
+    response = client.get(url, headers={"Authorization":"Bearer " + access_token})
     
     assert response.status_code == status.HTTP_200_OK
     assert response.json()[0]['title'] == todo.title
@@ -70,6 +71,8 @@ def test_todo_create(db, client):
     Test that a new object is successfully created when POST request is made to '/api/todos' successfully
     """
     
+    # login user
+    access_token, refresh_token = auto_login_user()
     # should begin with zero Todo objects
     assert Todo.objects.count() == 0
     # define model values
@@ -83,7 +86,9 @@ def test_todo_create(db, client):
                             {
                                 'title':title, 
                                 "memo":memo,
-                            })
+                            },
+                            headers={"Authorization":"Bearer " + access_token}
+                        )
                             
     assert response.status_code == status.HTTP_201_CREATED
     assert Todo.objects.count() == 1
@@ -94,11 +99,13 @@ def test_todo_detail(db, client, create_todo):
     Test that a get request can be made to '/api/todos/<id>' to obtain the detail page of a todo object successfully
     """
 
+    # login user
+    access_token, refresh_token = auto_login_user()
     # create todo object in db
     todo = create_todo(title='Learn how to use pytest', memo="Use the book 'Python Testing with Pytest'")
     # get url for retrieving detail page for specific blog
     url = reverse('todo-detail', args=(todo.pk,))
-    response = client.get(url)
+    response = client.get(url, headers={"Authorization":"Bearer " + access_token})
 
     assert response.status_code == status.HTTP_200_OK
     assert response.json()['id'] == todo.pk
@@ -109,6 +116,8 @@ def test_todo_update(db, client, create_todo):
     Test that a PUT request can be made to '/api/todos/<id>' to update a todo object successfully
     """
 
+    # login user
+    access_token, refresh_token = auto_login_user()
     # create todo object in db
     todo = create_todo(title='Learn how to use pytest', memo="Use the book 'Python Testing with Pytest'")
     # set values to be updated
@@ -122,6 +131,7 @@ def test_todo_update(db, client, create_todo):
                             'title':new_title,
                             'memo':new_memo,
                         },
+                        headers={"Authorization":"Bearer " + access_token},
                         content_type="application/json")
     assert response.status_code == status.HTTP_200_OK
 
@@ -135,6 +145,8 @@ def test_todo_delete(db, client, create_todo):
     Test that a DELETE request can be made to '/api/todos/<id>' to destroy a todo object successfully
     """
 
+    # login user
+    access_token, refresh_token = auto_login_user()
     # create todo object in db
     todo = create_todo(title='Learn how to use pytest', memo="Use the book 'Python Testing with Pytest'")
     # assert there are exactly 1 Todo objects
@@ -143,7 +155,7 @@ def test_todo_delete(db, client, create_todo):
     # get url for updating the todo object
     url = reverse('todo-detail', args=(todo.pk,))
     # make delete request to '/api/todos/<id>'
-    response = client.delete(url)
+    response = client.delete(url, headers={"Authorization":"Bearer " + access_token})
     
     assert response.status_code == status.HTTP_204_NO_CONTENT
     assert Todo.objects.count() == 0
